refactor(operations): convert tools module to execute-form (pilot)
First module authored transport-neutrally: all 8 tool-analytics operations
move into @repo/operations as execute(input, ctx) Effects with declarative
group/sdkMethod and rateLimitTier, mounted via mountOperationModules. The
new typedResponses helper (literal-keyed twin of openApiResponses, same
runtime) drives OperationOutput inference; keys stay literal — deriving
them from a generic extraErrors param collapses to a number index under
tsgo, so extraErrors statuses stay untyped for now.

openapi.json/mcp.json byte-identical; the pre-existing HTTP route tests
pass unchanged against the generated handler.
