--- apps/api/src/openapi/entities/tool.ts
@@ -1,2 +1,0 @@
-// Transitional shim while route modules migrate into @repo/operations.
-export * from "@repo/operations/openapi/entities/tool"
